@@ -17,16 +17,18 @@
 package com.example.android.teatime;
 
 
-import static android.support.test.espresso.Espresso.onData;
-import static android.support.test.espresso.action.ViewActions.click;
-import static android.support.test.espresso.matcher.ViewMatchers.withId;
+import static androidx.test.espresso.Espresso.onData;
+import static androidx.test.espresso.action.ViewActions.click;
+import static androidx.test.espresso.matcher.ViewMatchers.withId;
 
 import static org.hamcrest.Matchers.anything;
 
-import android.support.test.espresso.Espresso;
-import android.support.test.espresso.IdlingResource;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
+import androidx.test.core.app.ActivityScenario;
+import androidx.test.espresso.Espresso;
+import androidx.test.espresso.IdlingRegistry;
+import androidx.test.espresso.IdlingResource;
+import androidx.test.ext.junit.rules.ActivityScenarioRule;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.After;
 import org.junit.Before;
@@ -63,8 +65,8 @@ public class IdlingResourceMenuActivityTest {
      * complete. This rule allows you to directly access the activity during the test.
      */
     @Rule
-    public ActivityTestRule<MenuActivity> mActivityTestRule =
-            new ActivityTestRule<>(MenuActivity.class);
+    public ActivityScenario<MenuActivity> mActivityScenario =
+            ActivityScenario.launch(MenuActivity.class);
 
     private IdlingResource mIdlingResource;
 
@@ -72,9 +74,11 @@ public class IdlingResourceMenuActivityTest {
     // Registers any resource that needs to be synchronized with Espresso before the test is run.
     @Before
     public void registerIdlingResource() {
-        mIdlingResource = mActivityTestRule.getActivity().getIdlingResource();
-        // To prove that the test fails, omit this call:
-        Espresso.registerIdlingResources(mIdlingResource);
+        mActivityScenario.onActivity(activity -> {
+            mIdlingResource = activity.getIdlingResource();
+            // To prove that the test fails, omit this call:
+            IdlingRegistry.getInstance().register(mIdlingResource);
+        });
     }
 
     @Test
@@ -86,7 +90,7 @@ public class IdlingResourceMenuActivityTest {
     @After
     public void unregisterIdlingResource() {
         if (mIdlingResource != null) {
-            Espresso.unregisterIdlingResources(mIdlingResource);
+            IdlingRegistry.getInstance().unregister(mIdlingResource);
         }
     }
 }
