@@ -18,17 +18,17 @@ package com.example.android.teatime;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.annotation.VisibleForTesting;
-import android.support.test.espresso.IdlingResource;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.AdapterView;
-import android.widget.GridView;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.annotation.VisibleForTesting;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.test.espresso.IdlingResource;
 
 import com.example.android.teatime.IdlingResource.SimpleIdlingResource;
+import com.example.android.teatime.databinding.ActivityMenuBinding;
 import com.example.android.teatime.model.Tea;
 
 import java.util.ArrayList;
@@ -36,6 +36,7 @@ import java.util.ArrayList;
 public class MenuActivity extends AppCompatActivity implements ImageDownloader.DelayerCallback {
 
     Intent mTeaIntent;
+    private ActivityMenuBinding binding;
 
     public final static String EXTRA_TEA_NAME = "com.example.android.teatime.EXTRA_TEA_NAME";
 
@@ -58,10 +59,10 @@ public class MenuActivity extends AppCompatActivity implements ImageDownloader.D
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_menu);
-        Toolbar menuToolbar = (Toolbar) findViewById(R.id.menu_toolbar);
-        setSupportActionBar(menuToolbar);
-        getSupportActionBar().setTitle(getString(R.string.menu_title));
+        binding = ActivityMenuBinding.inflate(getLayoutInflater());
+        setContentView(binding.getRoot());
+        setSupportActionBar(binding.menuToolbar);
+        if (getSupportActionBar() != null) getSupportActionBar().setTitle(getString(R.string.menu_title));
 
         // Get the IdlingResource instance
         getIdlingResource();
@@ -69,7 +70,7 @@ public class MenuActivity extends AppCompatActivity implements ImageDownloader.D
 
     /**
      * We call ImageDownloader.downloadImage from onStart or onResume instead of in onCreate
-     * to ensure there is enougth time to register IdlingResource if the download is done
+     * to ensure there is enough time to register IdlingResource if the download is done
      * too early (i.e. in onCreate)
      */
     @Override
@@ -85,17 +86,16 @@ public class MenuActivity extends AppCompatActivity implements ImageDownloader.D
     @Override
     public void onDone(ArrayList<Tea> teas) {
 
-        //TextView testing =(TextView)findViewById(R.id.textView);
-        //testing.setText("Changed");
+        //TextView testing =findViewById(R.id.textView);
+        //binding.textView.setText("Changed");
 
         // Create a {@link TeaAdapter}, whose data source is a list of {@link Tea}s.
         // The adapter know how to create grid items for each item in the list.
-        GridView gridview = (GridView) findViewById(R.id.tea_grid_view);
         TeaMenuAdapter adapter = new TeaMenuAdapter(this, R.layout.grid_item_layout, teas);
-        gridview.setAdapter(adapter);
+        binding.teaGridView.setAdapter(adapter);
 
         // Set a click listener on that View
-        gridview.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+        binding.teaGridView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             @Override
             public void onItemClick(AdapterView<?> adapterView, View view, int position, long id) {
 
