@@ -18,22 +18,27 @@ package com.example.android.teatime;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
-import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.android.teatime.databinding.ActivityOrderSummaryBinding;
 
 import java.text.NumberFormat;
+import java.util.Locale;
 
 public class OrderSummaryActivity extends AppCompatActivity {
+
+    private ActivityOrderSummaryBinding binding;
+    private final Locale LOCALE = getResources().getConfiguration().getLocales().get(0);
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_order_summary);
-        Toolbar menuToolbar = (Toolbar) findViewById(R.id.order_summary_toolbar);
-        setSupportActionBar(menuToolbar);
-        getSupportActionBar().setTitle(getString(R.string.order_summary_title));
+        binding = ActivityOrderSummaryBinding.inflate(getLayoutInflater());
+        setContentView(binding.getRoot());
+        setSupportActionBar(binding.orderSummaryToolbar);
+        if (getSupportActionBar() != null) getSupportActionBar().setTitle(getString(R.string.order_summary_title));
 
         Intent intent = getIntent();
         String teaName = intent.getStringExtra(OrderActivity.EXTRA_TEA_NAME);
@@ -59,36 +64,23 @@ public class OrderSummaryActivity extends AppCompatActivity {
             String sugarType, int quantity) {
 
         // Set tea name in order summary
-        TextView teaNameTextView = (TextView) findViewById(
-                R.id.summary_tea_name);
-        teaNameTextView.setText(teaName);
+        binding.summaryTeaName.setText(teaName);
 
         // Set quantity in order summary
-        TextView quantityTextView = (TextView) findViewById(
-                R.id.summary_quantity);
-        quantityTextView.setText(String.valueOf(quantity));
+        binding.summaryQuantity.setText(String.valueOf(quantity));
 
         // Set tea size in order summary
-        TextView sizeTextView = (TextView) findViewById(
-                R.id.summary_tea_size);
-        sizeTextView.setText(size);
+        binding.summaryTeaSize.setText(size);
 
         // Set milk type in order summary
-        TextView milkTextView = (TextView) findViewById(
-                R.id.summary_milk_type);
-        milkTextView.setText(milkType);
+        binding.summaryMilkType.setText(milkType);
 
         // Set sugar amount in order summary
-        TextView sugarTextView = (TextView) findViewById(
-                R.id.summary_sugar_amount);
-        sugarTextView.setText(sugarType);
+        binding.summarySugarAmount.setText(sugarType);
 
         // Set total price in order summary
-        TextView priceTextView = (TextView) findViewById(
-                R.id.summary_total_price);
-
-        String convertPrice = NumberFormat.getCurrencyInstance().format(price);
-        priceTextView.setText(convertPrice);
+        String convertPrice = NumberFormat.getCurrencyInstance(LOCALE).format(price);
+        binding.summaryTotalPrice.setText(convertPrice);
 
     }
 
